requirements.txt
# AUTOLEADAI - Module 1: Call Handling and Speech Processing
# Requirements file

# Speech Recognition
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0

# Audio Processing
pyaudio==0.2.14
soundfile==0.12.1
librosa==0.10.1
noisereduce==3.0.0
pydub==0.25.1

# Machine Learning (Spam Detection)
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
joblib==1.3.2

# Utilities
python-dotenv==1.0.0
colorama==0.4.6

# Web Interface (Demo)
flask==3.0.0
flask-cors==4.0.0

# Database
sqlalchemy==2.0.23

